Impl Recipe & Fix Time to float

+++ src/de/fra/uas/digitales/kochbuch/backend/IRecipe.java
@@ -22,6 +22,6 @@ public interface IRecipe {
 
     int getRating();
 
-    int getTime();
+    float getTime();
 
 }
